// packages/client/src/components/Sections/SectionsList/constants.ts
export enum SECTIONS_LIST {
  TARGET_STANDARD = 'targetStandard',
  INDUSTRY = 'industry',
  ACTIVITIES = 'activities',
  PRODUCT_PAGES = 'productPages',
  COMPLEXITY = 'complexity',
  SPECIFIC_CONTENT = 'specificContent',
  INTERFACE_LANGUAGES = 'interfaceLanguages',
  CONFIGURATIONS = 'configurations',
  DELIVERABLES = 'deliverables',
  RATES = 'rates',
}